rename bqt.widget_manager to bqt.manager
